requirements.txt
streamlit
pyvista
numpy-stl
meshio
numpy